rename p2p/discovery/routing package, run go mod tidy
